backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
stripe==7.8.0
websockets==12.0
motor==3.3.2
pydantic==2.5.0
email-validator==2.1.0
bcrypt==4.1.2